Change API format
The `client` variable is a class property, not a static variable, so
change all the use sites for this object to the `@obj` syntax.

Also, all of these calls were directly to `request` which ought to be a
private method. `request_json` is the intended interface.

Performed with a `sed` so please inspect all the of the changes for
errors.